requirements.txt
Flask
gunicorn
openai
Pillow
numpy
opencv-python-headless
